Refactor constants
