+++ whispers/core/args.py
@@ -4,7 +4,8 @@ from functools import wraps
 from sys import argv, exit, stdout
 
 from whispers.__version__ import __version__, __whispers__
-from whispers.core.utils import DEFAULT_PATH, default_rules, load_regex
+from whispers.core.constants import DEFAULT_PATH
+from whispers.core.utils import default_rules, load_regex
 
 
 def argument_parser() -> ArgumentParser:
