+++ whispers/core/pairs.py
@@ -2,7 +2,8 @@ import logging
 from pathlib import Path
 from typing import Iterator, Optional
 
-from whispers.core.utils import REGEX_PRIVKEY_FILE, global_exception_handler, is_static, strip_string
+from whispers.core.constants import REGEX_PRIVKEY_FILE
+from whispers.core.utils import global_exception_handler, is_static, strip_string
 from whispers.models.appconfig import AppConfig
 from whispers.models.pair import KeyValuePair
 from whispers.plugins.config import Config
